Fix naming of request handlers in tchannel-example
*somebody* was overly aggressive with their class refactoring :)
