+++ tchannel-example/src/main/java/com/uber/tchannel/json/JsonServer.java
@@ -28,7 +28,7 @@ import io.netty.channel.ChannelFuture;
 public class JsonServer {
     public static void main(String[] args) throws Exception {
         final TChannel tchannel = new TChannel.Builder("json-server")
-                .register("json-endpoint", new JsonDefaultRequestHandler())
+                .register("json-endpoint", new JsonRequestHandler())
                 .setServerPort(8888)
                 .build();
