+++ tchannel-example/src/main/java/com/uber/tchannel/ping/PingRequestHandler.java
@@ -22,18 +22,23 @@
 
 package com.uber.tchannel.ping;
 
-import com.uber.tchannel.api.Request;
 import com.uber.tchannel.api.DefaultRequestHandler;
+import com.uber.tchannel.api.Request;
 import com.uber.tchannel.api.Response;
 import com.uber.tchannel.api.ResponseCode;
 
-public class PingDefaultRequestHandler extends DefaultRequestHandler<Ping, Pong> {
+public class PingRequestHandler extends DefaultRequestHandler<Ping, Pong> {
 
     @Override
     public Response<Pong> handle(Request<Ping> request) {
 
-        return new Response.Builder<>(new Pong("pong!"), request.getEndpoint(), ResponseCode.OK)
+        return new Response.Builder<>(
+                new Pong("pong!"),
+                request.getEndpoint(),
+                ResponseCode.OK
+        )
                 .setHeaders(request.getHeaders())
+                .setTransportHeaders(request.getTransportHeaders())
                 .build();
 
     }
