+++ tchannel-example/src/main/java/com/uber/tchannel/thrift/KeyValueServer.java
@@ -38,8 +38,8 @@ public class KeyValueServer {
         TChannel tchannel = new TChannel.Builder("keyvalue-service")
                 .setServerPort(8888)
                 .setLogLevel(LogLevel.INFO)
-                .register("KeyValue::getValue", new GetValueHandlerDefault(keyValueStore))
-                .register("KeyValue::setValue", new SetValueHandlerDefault(keyValueStore))
+                .register("KeyValue::getValue", new GetValueRequestHandler(keyValueStore))
+                .register("KeyValue::setValue", new SetValueRequestHandler(keyValueStore))
                 .build();
 
         tchannel.listen().channel().closeFuture().sync();
